Eigenfunction Decomposition of a (Transformed) Spatial Connectivity Matrix

Extract eigenvectors and corresponding eigenvalues from the matrix MWM, where M denotes a symmetric and idempotent projection matrix and W is the spatial connectivity matrix. This function also reports the Moran coefficient associated with each of the eigenvectors.

getEVs(W, covars = NULL)

Arguments

W

spatial connectivity matrix

covars

vector/ matrix of regressors included in the construction of the projection matrix M - see Details

Value

A list containing the following objects:

vectors

matrix of all eigenvectors

values

vector of the corresponding eigenvalues

moran

vector of the Moran coefficients associated with the eigenvectors

Details

The eigenfunctions obtained by getEVs can be used to perform supervised eigenvector selection and to manually create a spatial filter. To this end, a candidate set may be determined by 1) the sign of the spatial autocorrelation in model residuals and 2) the strength of spatial association found in each eigenvector as indicated by moran.

Prior to the spectral decomposition, getEVs symmetrizes the spatial connectivity matrix by: 1/2 * (W + W').

If covars are supplied, the function uses the covariates to construct projection matrix: M = I - X (X' X)^-1X'. Using this matrix results in a set of eigenvectors that are uncorrelated to each other as well as to the covariates. If covars = NULL, only the intercept term is used to construct M. See e.g., Griffith and Tiefelsdorf (2007) for more details on the appropriate choice of M.

References

Tiefelsdorf, Michael and Daniel A. Griffith (2007): Semiparametric filtering of spatial autocorrelation: the eigenvector approach. Environment and Planning A: Economy and Space, 39 (5): pp. 1193 - 1221.

See also

Author

Sebastian Juhl

Examples

data(fakedata)

E <- getEVs(W = W, covars = NULL)