Synthetic Dataset

An artificially generated cross-sectional dataset together with an accompanying binary connectivity matrix W. The \(n=100\) units are located on a regular grid and W is defined according to rook's adjacency definition of contiguity. The synthetic data can be used to illustrate the functionality of this package.

data(fakedata)

W

Format

An object of class data.frame with 100 rows and 8 columns.

An object of class matrix (inherits from array) with 100 rows and 100 columns.

Value

The file contains two objects:

fakedataset		a synthetic dataset

W		an artificial spatial connectivity matrix

Examples

data(fakedata)
head(fakedataset)
#>   ID indicator count       x1        x2         x3  negative negcount
#> 1  1         0     0 11.29514  5.527426  6.6406409 -2.338103        9
#> 2  2         0     4 13.92705  7.047870  4.0049350 -2.048737        3
#> 3  3         1     1 17.31431 10.525834 -0.7022595 -1.868379        5
#> 4  4         1     4 15.21778  7.817901  1.5948141 -2.737914        2
#> 5  5         1     3 15.56298  7.903947 -1.2708973 -1.117427        2
#> 6  6         1     2 17.43941 11.254842 -1.4158475  1.577442        2
dim(W)
#> [1] 100 100